Updated text messages

+++ app/handlers/rooms/create_room.py
@@ -20,7 +20,7 @@ async def create_room(message: types.Message):
                                       callback='cancel')
 
     await message.answer(
-        '"Хо-хо-хо! 🎅\n\n'
+        'Хо-хо-хо! 🎅\n\n'
         'Как ты хочешь назвать свою комнату?\n'
         'Напиши мне ее название и мы пойдем дальше\n\n'
         'Имя комнаты не должно превышать 12 символов.\n',
@@ -35,6 +35,17 @@ async def process_name(message: types.Message, state: FSMContext):
                                       callback='cancel')
     room_name = message.text
     await state.update_data(room_name=room_name)
+
+    if not len(room_name) < 13:
+        keyboard_inline = keyboard_button(text="Отмена",
+                                          callback='cancel')
+        return await message.reply(
+            text='Вы ввели слишком длинное имя, '
+                 'пожалуйста придумайте другое.\n'
+                 'Имя комнаты не должно превышать 12 символов.\n',
+            parse_mode=ParseMode.MARKDOWN,
+            reply_markup=keyboard_inline
+        )
 
     await CreateRoom.next()
     await bot.delete_message(chat_id=message.from_user.id,
